sistema descrizione

+++ personalPosts.php
@@ -27,7 +27,7 @@ include_once('Property.php');
                </div>
             </div>
 
-        <!-- Load table with annoncements if some have already been inserted -->
+        <!-- Load table with announcements if some have already been inserted -->
         <?php else: ?>
 
             <div class="col-md-9 posts-row">
@@ -50,8 +50,11 @@ include_once('Property.php');
                                         <th>Property ID</th>
                                         <th>Title</th>
                                         <th>Address</th>
-                                        <th>Date inserted</th>
-                                        <th>Days to expiration</th>
+                                        <th>Description</th>
+                                        <th><i class="fa fa-bed" style="font-size:25px;"></i></th>
+                                        <th><i class="fa fa-bath" style="font-size:25px;"></i></th>
+                                        <th><i class="fa fa-car" style="font-size:25px;"></i></th>
+                                        <th>Expiration</th>
                                         <th>Price</th>
                                         <th>Edit</th>
 
@@ -63,23 +66,41 @@ include_once('Property.php');
                                         $propertyArray = DBConnection::getPropertiesByUserId($_SESSION['userId']);
 
                                         foreach ($propertyArray as $property) {
-
                                             echo "<tr>";
-                                            echo "<td>" .  $property->getPropertyId() . "</td>";
-                                            echo "<td>" .  $property->getTitle() . "</td>";
-                                            echo "<td>" .  $property->getAddress() . "</td>";
-                                            echo "<td>" .  $property->getDateInserted() . "</td>";
-                                            echo "<td>" .  $property->getUpdateDate() . "</td>";
-                                            echo "<td>" .  $property->getPrice() . "</td>";
-                                            echo "<td> 
-                                                    <a href=\"#\" title=\"Extend announcement expiration date\"><i class=\"fa fa-calendar\"></i></a>";
-                                                    echo "<a href=\"#\" title=\"Edit announcement\"><i class=\"fa fa-pencil-square-o\"></i></a>
-                                                    <a href=\"propertyDeletion.php?delete=true&propertyId="; echo $property->getPropertyId(); echo "\" title=\"Delete announcement\"><i class=\"fa fa-trash\"></i></a>
+                                            echo "<td>" . $property->getPropertyId() . "</td>";
+                                            echo "<td>" . $property->getTitle() . "</td>";
+                                            echo "<td>" . $property->getAddress() . "</td>";
+                                            echo "<td>";
+                                            if (strlen($property->getDescription()) > 50) {
+                                                echo substr($property->getDescription(), 0, 50);
+                                                echo '...<a id="descriptionLoader"">&nbsp;[Read more]</a>';
+                                            } else {
+                                                echo $property->getDescription();
+                                            }
+                                            echo "</td>";
+                                            echo "<td>" . $property->getBedrooms() . "</td>";
+                                            echo "<td>" . $property->getBathrooms() . "</td>";
+                                            echo "<td>"; if ($property->getParking() == 0) {echo "No";} else {echo "Yes";}; echo "</td>";
+                                            echo "<td>" . $property->getUpdateDate() . "</td>";
+                                            echo "<td>" . $property->getPrice() . "</td>";
+                                            echo "<td><div class=\"edit-buttons-row\">";
+                                                    echo "<div class=\"calendar-button\"><a onclick=\"extendExpiration("; echo $property->getPropertyId(); echo ");\" title=\"Extend announcement expiration date\"><i class=\"fa fa-calendar\"></i></a></div>";
+                                                    echo "<div class=\"edit-button\"><a onclick=\"editProperty(";
+                                                    echo $property->getPropertyId();
+                                                    echo ")\" ";
+                                                    echo "\"  title=\"Edit announcement\"><i class=\"fa fa-pencil-square-o\"></i></a></div>
+                                                    <div id=\"delete-property-button\" class=\"delete-button\"><a onclick=\"return confirm('Are you sure?');\" href=\"propertyDeletion.php?delete=true&propertyId="; echo $property->getPropertyId(); echo "\" title=\"Delete announcement\"><i class=\"fa fa-trash\"></i></a></div></div>
                                                   </td>";
                                             echo "</tr>";
+
+                                            echo "<!-- The Modal --> <div id=\"myModal\" class=\"modal\">";
+                                            echo "<div class=\"modal-content\">";
+                                            echo "<div class=\"modal-header\">";
+                                            echo "<span id=\"close-description-modal\" class=\"close-modal\">&times;</span>";
+                                            echo $property->getDescription();
+                                            echo "</div></div></div>";
                                         }
                                     ?>
-
                                     </tbody>
                                 </table>
                             </div>
@@ -91,28 +112,28 @@ include_once('Property.php');
         <?php endif; ?>
 
         <div id="announcementForm">
-            <form class="form-container" method="post" action="userpage.php">
+            <form name = "propertyForm" class="form-container" method="post" action="userpage.php">
                 <h1>Insert details</h1>
 
                 <div>
                     <label for="title"><b>Title</b></label>
-                    <input type="text" placeholder="Title" name="title" maxlength="60" size="60" required>
+                    <input id="form-property-title" type="text" placeholder="Title" name="title" maxlength="60" size="60" required>
                 </div>
 
                 <div>
                     <label for="address"><b>Address</b></label>
-                    <input type="text" placeholder="Address" name="address" maxlength="60" size="60" required>
+                    <input id="form-property-address" type="text" placeholder="Address" name="address" maxlength="60" size="60" required>
                 </div>
 
                 <div>
                     <label for="description"><b>Description</b></label>
-                    <textarea class="form-control" rows="5" name="description" required></textarea>
+                    <textarea id="form-property-description" class="form-control" rows="5" name="description" required></textarea>
                 </div>
 
                 <div class="facilities">
                     <div>
                         <label for="bedrooms"><b><i class="fa fa-bed" style="font-size:30px;"></i></b></label>
-                        <select name="bedrooms" required>
+                        <select id="form-property-bedrooms" name="bedrooms" required>
                             <option value="1">1</option>
                             <option value="2">2</option>
                             <option value="3">3</option>
@@ -128,7 +149,7 @@ include_once('Property.php');
 
                     <div>
                         <label for="bathrooms"><b><i class="fa fa-bath" style="font-size:30px;"></i></b></label>
-                        <select name="bathrooms" required>
+                        <select id="form-property-bathrooms" name="bathrooms" required>
                             <option value="1">1</option>
                             <option value="2">2</option>
                             <option value="3">3</option>
@@ -144,7 +165,7 @@ include_once('Property.php');
 
                     <div>
                         <label for="parking"><b><i class="fa fa-car" style="font-size:30px;"></i></b></label>
-                        <select name="parking" required>
+                        <select id="form-property-parking" name="parking" required>
                             <option value="1">Yes</option>
                             <option value="0">No</option>
                         </select>
@@ -155,17 +176,17 @@ include_once('Property.php');
                 <div class="details">
                     <div>
                         <label for="area"><b>Area (in m^2)</b></label>
-                        <input type="number" placeholder="Area" name="area" min="1" max="10000" required>
+                        <input id="form-property-area" type="number" placeholder="Area" name="area" min="1" max="10000" required>
                     </div>
 
                     <div>
                         <label for="year"><b>Property Building Year</b></label>
-                        <input type="number" pattert="[0-9]{4}" placeholder="Year" name="year" min="1600" max="2019"required>
+                        <input id="form-property-year" type="number" pattert="[0-9]{4}" placeholder="Year" name="year" min="1600" max="2019"required>
                     </div>
 
                     <div>
                         <label for="price"><b>Price per month (in €)</b></label>
-                        <input type="number" pattert="[0-9]{4}" placeholder="Price" name="price" min="100" required>
+                        <input id="form-property-price" type="number" pattert="[0-9]{4}" placeholder="Price" name="price" min="100" required>
                     </div>
                 </div>
 
@@ -179,3 +200,25 @@ include_once('Property.php');
 
     </div>
 </div>
+
+<script>
+    // Get the modal
+    var modal = document.getElementById("myModal");
+
+    // When the user clicks the button, open the modal
+    document.getElementById("descriptionLoader").onclick = function() {
+        modal.style.display = "block";
+    }
+
+    // When the user clicks on <span> (x), close the modal
+    document.getElementById("close-description-modal").onclick = function() {
+        modal.style.display = "none";
+    }
+
+    // When the user clicks anywhere outside of the modal, close it
+    window.onclick = function(event) {
+        if (event.target == modal) {
+            modal.style.display = "none";
+        }
+    }
+</script>
